Adding no-kerf .dxf files
Adding no-kerf files, mx style opening holes with 7mm padding between the switch hole and edge of the keyboard. Use this style only for laser cut materials that are 1.5mm, or .06" thickness.
Base is the base. Stack plates are just perimeter plates for adding thickness to the keyboard. The stack plates are not needed unless you want to add them.
Fastener holes are 2mm diameter.