Logo I designed for my Pteron56
The font of the logo in my picture in /builds is Another America by The Branded Quotes.
Since that is a paid font I did not use that in this .svg. I used the font Edo by Vic Fieger available here 
https://www.fontsquirrel.com/fonts/Edo?q%5Bterm%5D=edo&q%5Bsearch_check%5D=Y

Also added is the .png version of the logo.